Revert "Importing."
This reverts commit d76218cf6c581eff02b7a3caf5dea35a6a237e45.

+++ src/main/java/com/asad/geniusanalysis/repository/ArtistRepository.java
@@ -2,7 +2,8 @@ package com.asad.geniusanalysis.repository;
 
 import com.asad.geniusanalysis.entity.Artist;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
-public interface ArtistRepository extends JpaRepository<Artist, Long> {
+public interface ArtistRepository extends JpaRepository<Artist, Integer> {
 
 }
